requirements.txt
# RTIE - Radiator Thermal Intelligence Engine
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & Feature Engineering
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# ONNX Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Explainability
grad-cam>=1.4.8

# Utilities
Pillow>=9.5.0
tqdm>=4.65.0
